Refactor github service
