+++ src/composables/useGithubUserQuery.ts
@@ -1,8 +1,8 @@
-import { getGithubApi, Github } from "../utils/github";
+import { getGithubService, Github } from "../utils/github";
 import { QueryKeys } from "../utils/QueryKeys";
 
 export default function (token: Ref<string | null>) {
-  const api = getGithubApi();
+  const api = getGithubService();
 
   return useQuery<Github.User | undefined>(
     [QueryKeys.GithubUser, token],
