+++ src/utils/github/api-cache.ts
@@ -1,0 +1,5 @@
+import { GithubApi } from "./api";
+
+export function createCachedGithubApi(api: GithubApi): GithubApi {
+  throw Error("TODO: createCachedGithubApi");
+}
